Update boarding.md
Note that Discourse UX flow is *terrible* if you get added as a moderator without 2FA already set up. It effectively blocks your account insisting you enable 2FA to continue (which is probably reasonable), but you have to "enter your existing password to continue" as the first step of this.

If you have been using OIDC federation ("login with github") until now, there is no local password to enter.  So the only option is a password reset email, which effectively falls back to a weaker security mechanism in order to try and bootstrap a stronger one.

So, let's add that as an explicit step in advance